* 修正获取URL数据封装的一处错误 * 语言数据添加新的内容 * 修正初始化采集器一个逻辑错误 * 修正采集器其他一些细节错误

+++ controller/coll-operate.go
@@ -233,6 +233,9 @@ func (this *CollOperate) GetLog(name string) string{
 
 //clear log
 func (this *CollOperate) ClearLog(){
+	if IsFile(this.log.lastSrc) == false{
+		return
+	}
 	err = WriteFile(this.log.lastSrc,[]byte(""))
 	if err != nil{
 		log.NewLog("",err)
